Changed HEIC interpereter
ImageIO was not working due to error: `Could not find a backend to open <image> with iomode ri`
The fix involves saving and rereading the image as a JPG, then deleting it.
This isn't very efficient as it has to store the file on disk. Looking into a better way of doing it.